pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "market_data"
version = "0.1.0"
description = "Extract market data from polygon and calculate technical indicators"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "polygon-api-client", 
    "python-dotenv",
    "pandas",
]

